- Many misc compilation fixes - Changed min/max macros to MIN/MAX to avoid conflict with c++ stl - Added gas/particle read/write timers to io_artio - Removed old platform files, updated some to reflect removal of fftw support - Fixed misc bugs in agn support

+++ src/base/index_hash.c
@@ -61,7 +61,7 @@ index_hash *index_hash_create( int size, int64_t max_key, int *remote_index, int
 	int oldj;
 	index_hash *hash;
 	int iter = 0;
-	int total;
+	size_t total;
 	int rekey;
 	index_hash_entry tmp_entry, new_entry;
